backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use[cli]>=0.5.9",
    "fastapi>=0.116.1",
    "hypercorn>=0.17.3",
    "livekit-api>=1.0.5",
    "livekit>=0.18.0",
    "livekit-agents>=0.8.0",
    "livekit-plugins-openai>=0.8.0",
    "livekit-plugins-deepgram>=0.8.0",
    "livekit-plugins-cartesia>=0.8.0",
    "livekit-plugins-turn-detector>=0.8.0",
    "livekit-plugins-silero>=0.8.0",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.6.1",
    "supabase>=2.18.1",
    "livekit-plugins-noise-cancellation~=0.2",
    "pydantic-ai-slim[mcp]>=0.7.4",
]
